requirements.txt
groq
mcp
mcp-simple-arxiv
python-dotenv